more renames and changes
