+++ .github/workflows/test.yml
@@ -2,7 +2,7 @@ name: Tests
 on:
   push:
     branches:
-      - jsign/retr2
+      - jsign/retr4
   pull_request:
     branches:
       - master
