+++ ffs/scheduler/internal/astore/astore.go
@@ -7,13 +7,10 @@ import (
 
 	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-datastore"
-	logging "github.com/ipfs/go-log/v2"
 	"github.com/textileio/powergate/ffs"
 )
 
 var (
-	log = logging.Logger("ffs-sched-astore")
-
 	// ErrNotFound indicates the instance doesn't exist.
 	ErrNotFound = errors.New("action not found")
